+++ kamehouse-vlcrc/src/main/java/com/nicobrest/kamehouse/vlcrc/config/VlcRcWebSocketConfig.java
@@ -26,10 +26,7 @@ public class VlcRcWebSocketConfig implements WebSocketMessageBrokerConfigurer {
   public void registerStompEndpoints(StompEndpointRegistry registry) {
     registry.addEndpoint("/api/ws/vlcrc/desktop")
         .setAllowedOriginPatterns("*");
-    registry.addEndpoint("/api/ws/vlc-player/status")
-        .setAllowedOriginPatterns("*")
-        .withSockJS();
-    registry.addEndpoint("/api/ws/vlc-player/playlist")
+    registry.addEndpoint("/api/ws/vlcrc/sockjs")
         .setAllowedOriginPatterns("*")
         .withSockJS();
   }