+++ kamehouse-vlcrc/src/test/java/com/nicobrest/kamehouse/vlcrc/integration/VlcRcPlaylistWebSocketControllerIntegrationTest.java
@@ -18,7 +18,7 @@ class VlcRcPlaylistWebSocketControllerIntegrationTest extends
 
   @Override
   public String getWebSocketUrl() {
-    return "/api/ws/vlc-player/playlist/";
+    return "/api/ws/vlcrc/sockjs/";
   }
 
   @Override