+++ kamehouse-vlcrc/src/test/java/com/nicobrest/kamehouse/vlcrc/integration/VlcRcStatusWebSocketControllerIntegrationTest.java
@@ -18,7 +18,7 @@ class VlcRcStatusWebSocketControllerIntegrationTest extends
 
   @Override
   public String getWebSocketUrl() {
-    return "/api/ws/vlc-player/status/";
+    return "/api/ws/vlcrc/sockjs/";
   }
 
   @Override